move gaggle test into one_taskset
